requirements.txt
flask>=2.0.1
requests>=2.25.1
python-dotenv>=0.19.0
flask-cors>=3.0.10
PyPDF2>=2.0.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-huggingface>=0.0.5
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
faiss-cpu>=1.7.3
sentence-transformers>=2.2.2
ollama>=0.1.5
python-multipart>=0.0.5
transformers>=4.30.0
diffusers>=0.21.4
torch>=2.0.0
accelerate>=0.20.3
psutil>=5.9.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0
itsdangerous>=2.0.0
Jinja2>=3.0.0
MarkupSafe>=2.0.0
Werkzeug>=2.0.0
click>=8.0.0
networkx
spacy
opencv-python
numpy
Pillow
gtts
moviepy
python-multipart
pylatex>=1.4.1
pandas>=1.3.0
matplotlib>=3.4.0
katex>=0.13.0
pandoc>=2.0.0